Extracted functions into another file and corrected the string
